Moved Participants List related functions to participants.py
